--- src/utils/request/index.js
@@ -1,2 +1,0 @@
-import request from './request';
-export default request;
